IO: make buffer release guaranteed for cases
Make byte channel release if it has been closed with
error or if write preparation loop has allocated a buffer
but failed to start writing (for example, joining started
so we have to cancel wriging preparation).

It is important to release allocated/borrowed buffer during
write operation preparation and if CAS loop was not successful
and no write state was set due to some reason (for example, joining).